pyproject.toml
[tool.poetry]
name = "parsing"
version = "0.1.0"
description = "empty string"
authors = ["kpomak <romka_ne@mail.ru>"]
license = "pirates"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.16"
requests = "^2.28.1"
lxml = "^4.9.2"
bs4 = "^0.0.1"
black = "^22.12.0"
pymongo = "^4.3.3"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
